minor changes #4 (help, im being held at gunpoint 🔫🔫🔫)

+++ src/events/MessageUpdate.ts
@@ -24,6 +24,8 @@ export default class MessageUpdateEvent extends BaseEvent {
 			oldMessage.guildId,
 		);
 
+		if (!guildPreferences) return;
+
 		if (oldMessage.channelId === guildPreferences.countingChannelId)
 			await newMessage.delete();
 	}
